Protecting Routes

+++ controllers/videoController.js
@@ -46,7 +46,10 @@ export const postUpload = async(req, res) => {
         posterFileUrl: posterPath,
         title,
         description,
+        creator: req.user.id
     });
+    req.user.videos.push(newVideo.id);
+    req.user.save();
     res.redirect(routes.videoDetail(newVideo.id));
 };
 
@@ -55,7 +58,7 @@ export const videoDetail = async(req, res) => {
         params: { id },
     } = req;
     try {
-        const video = await Video.findById(id);
+        const video = await Video.findById(id).populate("creator");
         res.render("videoDetail", { pageTitle: video.title, video });
     } catch (error) {
         res.redirect(routes.home);
@@ -69,10 +72,11 @@ export const getEditVideo = async(req, res) => {
 
     try {
         const video = await Video.findById(id);
-        res.render("editVideo", {
-            pageTitle: `Edit ${video.title}`,
-            video,
-        });
+        if (`${video.creator}` !== `${req.user.id}`) { // obj와 string을 비교하기 위해
+            throw Error();
+        } else {
+            res.render("editVideo", { pageTitle: `Edit ${video.title}`, video, });
+        }
     } catch (error) {
         res.redirect(routes.home);
     }
@@ -83,7 +87,6 @@ export const postEditVideo = async(req, res) => {
         params: { id },
         body: { title, description },
     } = req;
-
     try {
         await Video.findOneAndUpdate({ _id: id }, { title, description });
         res.redirect(routes.videoDetail(id));
@@ -96,9 +99,15 @@ export const deleteVideo = async(req, res) => {
     const {
         params: { id },
     } = req;
+
     try {
-        await Video.findOneAndRemove({ _id: id });
-        console.log("delete video successfuly!");
+        const video = await Video.findById(id);
+        if (`${video.creator}` !== `${req.user.id}`) {
+            throw Error();
+        } else {
+            await Video.findOneAndRemove({ _id: id });
+            console.log("Delete video successfuly!");
+        }
     } catch (error) {
         console.log(error);
     }
